Stop using indexes on uploadtrees
Summary:
Read the summary of D31935606 (https://github.com/facebookexperimental/eden/commit/82bc6440367c848697511d8f579b21862c795f55)

This is also about not using indexes anymore in requests. This is about the type `UploadTreeResponse`, which is very similar to `LookupResponse`, but not the same, as `token` is not optional.

We also don't actually use the returned index on client side, so the transition is easier.

Steps:
1. Remove client code that depends on `UploadTreeResponse.index` **(this diff)**
2. Remove `index` from `UploadTreeResponse`, to prevent bad usage in the future. (next diff) (needs to wait for (1) to push)

Reviewed By: StanislavGlebik

Differential Revision: D32001742

fbshipit-source-id: 72b3e25ea8cad6c55189b1f58a56c85cf7df2b90