test-amend-metaedit: drop obsstore use
Summary:
We do not use the obsstore. Drop it from the test.

I think the config was added when the mutationstore missed not have the "copy
markers" behavior. But that behavior was added by D22174991 (https://github.com/facebookexperimental/eden/commit/868c2b01080b01f189d0aba1b45583cfbfabacdd). So there is no
need to use the legacy obsstore.

Differential Revision: D31877897

fbshipit-source-id: b6c902f406897c02267e232fc8f928c5644bbabc