benchmark filesystem based xattr calls
Summary:
EdenFS NFS does not support filesystem getxattr calls. This means for now tools
that rely on sha1 will need to use thrift. To understand how big of a regression
this is I am benchmarking the difference between reading files via thrift and
filesystem.

We already have a benchmark for measuring thrift sha1 accesses, so I am
extending this to support xattrs through the filesystem as well.

Reviewed By: chadaustin

Differential Revision: D31161699

fbshipit-source-id: 9f6b1d98de18c9ca720cfa18c7094d322ae4161a