mononoke: don't depend on MySQL libs in OSS builds
Summary:
We now have a Rust library deep down that dependency tree that is depended on
by C++ (which ends up being no-oped because this is what the OSS mode does), so
this dependency path causes us to attempt to build a C++ library without its
companion Rust library and everything explodes.

I explained in more detail in
https://fb.workplace.com/groups/scm.mononoke/permalink/1568810016814901/ what
the problem is, but the TL;DR is the OSS mode was designed to build Rust
binaries so it does not work very well to build Python binaries.

Reviewed By: dtolnay

Differential Revision: D32147012

fbshipit-source-id: f05909d1ad672c8497ad67e94d57ce5684656675