inodes: add a method to canonicalize a path
Summary:
On Windows, the path returned to the PRJ_GET_PLACEHOLDER_INFO_CB callback must
be canonical and may differ from the passed in path. Currently, this callback
is implemented by walking the Inode hierarchy, but this is being changed to
walk the Tree hierarchy. In the case of Inode, the getPath method can be used,
not for trees. Therefore let's add a method that walks the Tree hierarchy and
returns the proper cased path.

Reviewed By: chadaustin

Differential Revision: D32022637

fbshipit-source-id: fbf14ef806d3ca0d356a0a4d5374d4a48663d005