inodes: add EdenMount::getTreeOrTreeEntry method
Summary:
On Windows, the expected behavior from ProjectedFS callbacks is that they query
the "backing store" state. On EdenFS, we have so far assumed that the "backing
store" was EdenFS Inodes, but EdenFS's backing store state is actually what
needs querying. In order to do so, the ProjectedFS dispatcher must be able to
query the backing store from a path, which is what getTreeOrTreeEntry does.

This is similar to the EdenMount::getInode but acting on the backing store
Tree.

Reviewed By: chadaustin

Differential Revision: D32022636

fbshipit-source-id: df1dfe09740dcf3ab36d318a850e9e7e4942be60