lock: replace util.islocked with lock.islocked
Summary:
util.islocked function operated directly on a file name, which is peeking too far into the locking abstraction. Instead, replace with new function lock.islocked that uses the normal locking API to determine if a lock is taken. This minor change makes it easier to switch to a rust lock implementation.

I had to reintroduce a flag to skip the deadlock check since islocked tries to acquire a lock that might already be held (on purpose).

One existing user of util.islocked did not have a vfs object available, so I just removed it since it was only used for some debugging output.

Reviewed By: quark-zju

Differential Revision: D31963496

fbshipit-source-id: 67e68909519a789ed9b4f20f60d30c090ee8d3f8