lock: add high level rust lock
Summary: lock.rustlock is similar in behavior to lock.lock except that it uses rust locks under the hood. Its _trylock is somewhat simpler since the existing lock contents are returned on lock acquisition failure, so it is "not possible" for it to fail reading the existing lock contents when the lock is contended.

Reviewed By: quark-zju

Differential Revision: D31963493

fbshipit-source-id: 22e84525967ecabe0e970cf2ab7c77edda5e7a1c