Support creating a bubble with custom duration
Summary:
This diff allows creating an ephemeral bubble with a custom lifetime. This will be used when we need to create a snapshot with a custom retention.

For now, all callers of `create_bubble` still provide `None` as custom retention, meaning none of them have custom retention and current behaviour continues the same.

I also used `Duration` instead of `ChronoDuration` on the API as that is neater, in stdlib, and doesn't support negative durations (which we don't want anyway).

Reviewed By: StanislavGlebik

Differential Revision: D32103887

fbshipit-source-id: 484f018f1f13a74aca5c012fa50efe51f551ad05