dag: add IdDagStore::all_ids_in_segment_level
Summary:
Calculate IdSet covered by a specific level of segments.

The high-level segments are designed to be "missing the last segment per
level", like:

  Level 0 (flat) [-][-][-][-][-][-][-]  gap  [-][-][-][-][-][-][-][-]
  Level 1        [----][----][----]          [----][----][----]
  Level 2        [----------]                [----------]

This method will be used to figure out what lower level segments should
be checked to build higher level segments. Previously it was using the
`next_free_id` API, which does not work with discontinuous segments.

Reviewed By: DurhamG

Differential Revision: D32084971

fbshipit-source-id: 17d3fb410f963edbf52803c2d90464caf62e6afe