Don't require write context for snapshot create command
Summary:
There are permissions in Mononoke server. Depending on the exact functions you use to create a changeset, you might get a write permission, or just a read one.

This inconsistency cause stuff like commitcloud to work in Sandcastle, but not snapshots, even when doing the same operation (creating a commit).

This diff makes snapshot not require write permissions, so it works in Sandcastle.

The long-term path here is to reduce the current write permissions so they only gate modifying public data, and create a new permission to modify draft data, for stuff like commitcloud and snapshots.

Reviewed By: markbt

Differential Revision: D32253549

fbshipit-source-id: 05b9209a84e20f9adce2ab083bcba407b686dd80