mononoke: start removal of CommitSyncDataProvider
Summary:
CommitSyncDataProvider is a wrapper around LiveCommitSyncConfig. It was added
~1 year ago when a new way of rewriting commits was introduced.
The reason it was added was to preserve some of existing tests. These tests
used Movers/BookmarkRenamer that were impossible to express using our
LiveCommitSyncConfig, and so we created CommitSyncDataProvider as a wrapper
around LiveCommitSyncConfig so that it can preserve these tests.

However now CommitSyncDataProvider gets in a way of another initiative (see
D32206082). Since CommitSyncDataProvider was a hack to start with I'd like to
try and remove it in this stack of diffs. For the tests where it's possible to
do so I'm going to use TestLiveCommitSyncConfig, and for tests where it's not
possible I'm going to just delete them.

This stack of diff starts with replacing CommitSyncDataProvider::test_new with
CommitSyncDataProvier::Live in test_utils.

Reviewed By: yancouto

Differential Revision: D32242285

fbshipit-source-id: 31c22f1e6fd25db8461fe4b6b5afc4bce8e304a8