mononoke: add large repo commit version mapping
Summary:
This goal of this stack of diffs is to simplify adding new small repos to
an existing large repo. See attached task for more details.

In order to do that we need to add a new sql table that keeps track of a large
version for a given large repo commit. This diff starts using this table for
writing, but not yet for reading. This will follow in the next diffs.

Reviewed By: mitrandir77

Differential Revision: D32206082

fbshipit-source-id: e4cc1fbb470064e9b933f18ff87993f7b09470cd