prjfs: detach callbacks futures when not ready
Summary:
The QueuedImmediateExecutor can have surprising effect, let's instead move the
futures to the global CPU executor, but only if the ImmediateFuture isn't
ready, ie: when there is still some computation to be performed. The use of
folly::detachOnGlobalCPUExecutor also makes it obvious that the intent is to
run the future elsewhere in the background.

Reviewed By: chadaustin

Differential Revision: D32155651

fbshipit-source-id: 8ae5be48230ff44c58a5697d3824825eeb8235dc