sparse: remove caching
Summary:
In the old days computing sparse matchers was expensive because it
involved opening the flat manifest. Therefore we had a layer of caching. In a
future diff we allow sparse matchers to be influenced by current configs, which
makes it much harder to determine an appropriate cache key for the inputs
without just reading everything. Since reading the sparse profile is so much
cheaper these days, let's just get rid of this caching entirely.

Reviewed By: quark-zju

Differential Revision: D31705828

fbshipit-source-id: 551e863b31ce3b56b37bf1fd7bdddb4b902a45bd