template: support fast "revset(...)|count"
Summary:
Counting commits in revsets like `_firstancestors(x)` (or just `::x`) can be
interesting as a auto-incremental number for a linear (p1) branch.
The template syntax can express it but the implementation expands the revset,
making it very slow.

Update the template code to be lazy about revset so the counting can be fast.

This will be used by ngavalas's work to use O(1)-ish data to track a range of
commits.

See also https://fb.workplace.com/groups/sourcecontrolteam/posts/3616819875105999/,
which provides ways to translate between node and number:

  # hg dbsh
  branch = repo.dageval(lambda: firstancestors(lookup("master")))
  x2i = lambda x: len(repo.dageval(lambda: branch & ancestors(lookup(x))))
  i2x = lambda i: branch.skip(len(branch) - i).first()

  In [7]: hex(i2x(100))
  Out[7]: 'f8295905daa86e19bc71aa1f2de0e44242368960'

  In [8]: x2i('f8295905daa86e19bc71aa1f2de0e44242368960')
  Out[8]: 100

Reviewed By: DurhamG

Differential Revision: D32300766

fbshipit-source-id: b9dcb511344f44af0da798167ed21562d310f6dd