scmstore: change fetch result to be an iterator
Summary:
Now that the fetch results are produced via a background thread, let's
start returning the result channel to the caller so they can receive the results
as they're produced. This will help us consume the results and get them out of
memory so we don't pend the entire result in memory.

Reviewed By: quark-zju

Differential Revision: D32212908

fbshipit-source-id: 12a82990ac7002d8da0376a9b689a4ddfea49c27