scmstore: move aux_data writing to happen earlier
Summary:
Previously we stored all aux_data in memory, then wrote it to disk at
the end. In a future diff, we'll make it so we stream the fetch results out to
the caller as we get them, so we don't want to store things in memory for the
entire fetch.

To help with this, let's write aux_data to disk as soon as we've completed it.

Reviewed By: quark-zju

Differential Revision: D32212907

fbshipit-source-id: 9916604e832edd7f1ebe4d1e3bcdead61ab57807