scmstore: move tree indexedlog/memcache writing to be earlier
Summary:
In a future diff we'll want to stream the results of file/tree fetching
out as we receive stuff. This means we don't want to keep things in memory for
the lifetime of the fetch. For files, we already write stuff to disk as we
receive it, but let's do the same for trees now.

Reviewed By: quark-zju

Differential Revision: D32212911

fbshipit-source-id: fc0c5e0e5968b7625452af8fb5a6e998d52d7144