scmstore: remove metrics from fetch result object
Summary:
This isn't actually used anywhere. In a future diff we'll change the
fetch results to be a stream of results/errors and the metrics concept won't fit
in anymore. So let's remove it.

Metrics are stil collected on the file store object. They just aren't reported
for individual requests now.

Reviewed By: quark-zju

Differential Revision: D32212910

fbshipit-source-id: 9b747a842b0f217cfc9d3160134434b1e1d89cbb