scmstore: move fetching logic to a background thread
Summary:
Currently fetching runs synchronously on the main thread, which means
all the results are kept in memory and returned at the end. Ideally we'd stream
the results out of the fetch logic as we receive them, so they can be consumed
then discarded and not kept in memory.

Ideally we'd do this via async, but it ends up being quite complicated. So
instead let's just run the fetch on a thread and send the results to the mean
thread's iterator via a channel.

This first diff just moves the computation to a background thread. It still
consumes the results all at once before returning the results to the caller. A
future diff will change the return value to return the stream of results to the
caller.

Reviewed By: quark-zju

Differential Revision: D32212906

fbshipit-source-id: 1c7e5258cafd6be4405e2c5ce2038a716305a3cd