cli_rs: Remove crossterm refreshing
Summary:
This is because the save / restore cursor position sometimes behaves differently depending on the command that I run minitop with (e.g., it sometimes will re-print the first row repeatedly if the saved cursor position was somehow wrong).

I also explored calculating how much to truncate by getting the screen size, but this was simpler.

We can add this back in if we switch to an "alternate screen".

Reviewed By: fanzeyi

Differential Revision: D32289941

fbshipit-source-id: 3edda78ab2aa82958f6a40efd03e302520f84f4b