dag: move master/non-master option from 2 parameters to VertexOptions
Summary:
Previously we use two parameters to specify heads in the master group and
non-master group. Now we have the `VertexOptions` to specify additional
attributes. Let's merge 2 parameters and use the option to specify the group
so the APIs look more concise.

This changes many places because `flush(master_heads)` is changed to
`flush(heads)`. An extra step is needed to mark those `heads` as
`master_heads`.

Reviewed By: DurhamG

Differential Revision: D32272000

fbshipit-source-id: df92914f2ca8662c0aab134fb6d05e5bc0dfb232