RFC - scmquery/scsc: address unexpected diff behavior with large moved files
Summary:
# Context

I noticed some weird behavior with SCMQuery proxy and (large) moved files where:
- passing in just the destination path got us a "diffing too many bytes" error
- passing in just the origin path got us nothing
- passing in both paths got us the expected diff:
```
 diff --git a/xplat/arfx/testing/assets/SnapshotTests/blocks_nesting_prefabs.arexport b/xplat/arfx/testing/assets/UnifiedSnapshotTests/blocks_nesting_prefabs.arexport
rename from xplat/arfx/testing/assets/SnapshotTests/blocks_nesting_prefabs.arexport
rename to xplat/arfx/testing/assets/UnifiedSnapshotTests/blocks_nesting_prefabs.arexport
```

In the scmquery proxy it appears path pairs with a large enough size will have the 'generate_placeholder_diff' flag set. There is some logic that won't count the size of a file of a path if the path isn't in the input list of paths. This does not appear to account for the case where base_path and other_path are different as the result of a move or copy.

Noticed this issue also seemed to reproduce in `scsc diff`, but would error even if the origin path was passed in. I tried fixing that in this diff as well, even though it turned out to be unrelated to the case of moved files. The utility will only request a placeholder diff if explicitly asked from the command line arguments. It seems that this would repro with any sufficiently large changed file.

# Solution

SCMQuery Proxy - Always count the file size of both base_file and other_file in the pre-request file size check, as SCS will take both into account regardless

scsc - Set `generate_placeholder_diff` if `pair_size` is above the limit.

Reviewed By: markbt

Differential Revision: D32325673

fbshipit-source-id: 88c149726d4f1e92bb11a24203f8a8d642e08354