mononoke: remove SyncData from backsyncer
Summary:
This diff removes SyncData from backsyncer and uses
TestLiveCommitSyncConfig instead.

There's one caveat that's worth mentioning - there were a few tests that were
using movers that are impossible to create with our configs. For example, a
mover that strips a prefix from a small repo files or a bookmark renamer that
strips a prefix from small repo bookmarks. Since now we won't be able
to create movers like that this diff removes them.

Reviewed By: Croohand

Differential Revision: D32280414

fbshipit-source-id: b1796305228dec069971fb595f6efff46ae7bfa0