Fix sharding logic for time shards
Summary:
I was rolling out edenapi uploads and realised they jumped up quicker than I expected. I went to take a look, and it turns out a lot of our migrations are happening incorrectly. Not that much, though.

When generating the config, we generate a shard (a 0-99 number), and use that number for all sharding operations: percentage rollouts, timed rollouts, and group decision. The problem is it means those three migrations are always correlated.

This diff fixes that by using independent shard numbers for percentage migrations and time migrations.

Reviewed By: DurhamG

Differential Revision: D32434107

fbshipit-source-id: e5a44969aab421b9c430999c633d58f5ba19191a