runlog: disable runlog when running as root
Summary: The hg runlog causes every hg command to write out a runlog file (and create the runlog directory if it doesn't exist). If hg is run as root this can cause permission issues for subsequent hg commands. To mitigate this issue, the runlog feature now disables itself on unix platforms if the process's effective user id is 0 and the .hg directory is _not_ owned by root. The presumption is that if .hg is owned by root then it is normal for hg commands to be run as root.

Reviewed By: quark-zju

Differential Revision: D32368400

fbshipit-source-id: e6cbeb37e28a414a705b7a4eb350e87e34f6faf5