drawdag: use a single transaction
Summary:
This reduces disk flushes, and is more efficient.

The main motivation is to avoid a situation where each added commit becomes
part of `visibleheads`. For example, drawdag:

  C
  |
  B
  |
  A

would previously have visibleheads change from `A` to `B` to `C`, but now
`visibleheads` is only `C` without having the `B` or `A` state.

Reviewed By: singhsrb

Differential Revision: D32436069

fbshipit-source-id: 30274b1158c5eb796bc24101dc892c334c0f0287