integration: speedup hg test by 2x
Summary:
In every test setup, writing a file to the repository will always call `hg add`
in order to this newly added file to be added in subsequent `hg commit`. Since
`hg add` can be very expensive (in particular on Windows), some tests are
spending most of their time calling `hg add` instead of running the body of the
test. As an easy way to avoid this, let's stage the added files in memory
first, and then call `hg add` once just before `hg commit`.

Reviewed By: genevievehelsel

Differential Revision: D32333775

fbshipit-source-id: 8cc6cfad7ec73f6f0c50d7d1e6affa6cdd828e2f