No need to specify bubble in scsc command
Summary: Snapshots already have a bit of `scsc` support, but you needed to specify `--bubble-id` argument, which is not really good as that is not visible to users. This diff makes it possible to query a snapshot simply by using `--snapshot-id`. The bubble can still be provided to make the query faster.

Reviewed By: markbt

Differential Revision: D32395199

fbshipit-source-id: 8f89ae86b56d3789a5cbf69b6a4d98979e79ec4d