mononoke: make version mapping non-optional in WorkingCopyEquivalence
Summary:
In the next diff we are going to refactor CommitSyncOutcome::NotSyncCandidate
to have a version, which would mean that the version should always be present
even for NotSyncCandidate

So in this diff let's make mapping version non-optional. Note that at the moment our
prod tables have some of the mapping entries with a non-empty version - I'm
going to fix it before landing this diff.

Reviewed By: mitrandir77

Differential Revision: D32428740

fbshipit-source-id: be9f8b104a14497bf5a524c68c0fd53beba22ee1