inodes: add TreeInode::removeRecursively
Summary:
On Windows, directory removal notifications can be received out of order, and
sometimes duplicated. The next diff will rework the way EdenFS handle
notifications on Windows by looking at the on-disk state and applying it to the
inode hierarchy. In some cases, an entire directory hierarchy will need to be
removed which the added function will make it easier to do.

Reviewed By: chadaustin

Differential Revision: D32329806

fbshipit-source-id: ad9af389c993420184c0e59dd3d25188adddb77a