sparse: reintroduce in-memory caching for sparse matcher
Summary:
It turns out repo.sparsematch() is called in a tight loop in some
places, like during checkout's record update phase. Let's cache it in memory.

This is a little different from the old caching mechanism because 1) it's only
ever in memory, so we don't need to worry about persisting it over time and
invalidating it when the sparse profile configs change, and 2) it only hashes
the inputs, and not the whole config file. The one input that is not part of the
cache key is the dynamic sparse profile config. Since that input has strict
invalidation points, we can just invalidate the sparse cache when the dynamic
sparse profile config invalidated.

Differential Revision: D32474918

fbshipit-source-id: 35593d06dac420e7a7fee47c5e1602413541de5f