mononoke: use new value_len field from sqlblob chunk tables
Summary:
Add support for nullable value_len fields to sqlblob chunk tables and population of them from sqlblob_gc mark phase.

For put() this change adds a chunk_generation insert if the UpdateGeneration on put shows no rows affected. This is so that that chunk_generation.value_len can be relied on to be present for new rows. The insert is done after chunk and data put so as to not introduce a "chunk_generation without data key" state.

Reviewed By: farnz

Differential Revision: D32317235

fbshipit-source-id: 83de98350a59aab6905a7e6cf9f64d8f89c7682e