Don't redownload files if possible
Summary:
This is an optimisation for snapshots. If we're restoring an snapshot, and the files are already downloaded with correct content, we don't need to delete and redownload them.

This is useful for "syncing" mechanisms that quickly create and restore snapshots, as most of the files will be the same between restores, and thus don't need redownloading.

Reviewed By: StanislavGlebik

Differential Revision: D32288548

fbshipit-source-id: 6572c3d49584d350f06c2f8e2a3ac57e7c8696c9