Allow copying from bubble on lookup
Summary:
This diff allows specifying a `copy_from_bubble` on a lookup request. If the file is not found in your usual bubble during a request, it will check the `copy_from` bubble, and if the file is available there, will copy its content, so it doesn't need to be reuploaded.

This will be used to optimise uploads on snapshots, which often share a lot of files.

Reviewed By: markbt

Differential Revision: D32471433

fbshipit-source-id: 6c913ae79fbf19f8162d4faac2bad216294f6fb9