mononoke: fix how we decide if a version is related or not
Summary:
Previously we've been just looking at the small repos in a given version to
decide if this is related or not. This is wrong - versions for all pre-merge fbsource
commits have no small repos in it, and yet we want to find this versions when
trying to sync a commit from fbsource to ovrsource.

FWIW just looking at repos in given version is not enough to understand if a
repo and version are related. What tells them if they are related is common
config - if common config contains a given repo, then it's indeed related.

This is what this diff does.
Also I had to debug some of the failing tests, and pretty_assertions made them easier to debug, so I decided to leave pretty_assertions here.

Reviewed By: mitrandir77

Differential Revision: D32461769

fbshipit-source-id: 9c9a57aa71e20cec8448602dfbe0248a6bb850f7