Implement `dag` crate traits for server-side IdMap
Summary:
Used in the tailer to prove that they work - this is currently a "one commit at a time" situation, but will be optimized later.

The existing server code has its own system for assigning `IdMap` `DagId`s to `ChangesetId`s, intertwined with the `dag` crate's system for converting `DagId`s to segments. It also has no support for abbreviated `ChangesetId`s, required for using the `dag` crate.

Rather than completely rewrite the server-side logic, which has integration with Mononoke's existing monitoring and logging, create a wrapper struct that allows you to use the server `IdMap` trait with the `dag` crate `IdMap` algorithms; we already use the `dag` crate `IdDag` algorithms to create an `IdDag`, so this unifies behaviour with the client.

In our current use cases, abbreviated `ChangesetId` support is not used; it's implemented because an implementation is required by the `dag` crate, but the implementation may not be correct for all use cases.

With the required `dag` crate traits implemented on `IdMapWrapper`, the core of the change is in `fn once()` - instead of calling `build_incremental` to assign IDs in the `IdMap` and `IdDag` at once, we call `assign_head` from the `dag` crate to assign IDs, and use its return value to drive `iddag.build_segments_volatile_from_prepared_flat_segments`.

`build_segments_volatile_from_prepared_flat_segments`, in turn, takes the result of `assign_heads` and uses it to incrementally update the `IdDag`.

Reviewed By: quark-zju

Differential Revision: D32262361

fbshipit-source-id: 95801dbb981bb021dcde16bfeca2f353aca0b752