Add a ChangesetFetcher that uses bulk prefetched commits for speed
Summary:
In several places, notably in the segmented changelog tailer and seeder, we have an entity which uses a `ChangesetFetcher` to fetch changesets, but also wants to be able to run with bulk prefetched commits (via deserializing a file on disk, or via `bulkops`).

Create a `ChangesetFetcher` that keeps an in-memory prefetch set it can use to answer queries quickly.

Reviewed By: StanislavGlebik

Differential Revision: D32330206

fbshipit-source-id: 93dfb5c66121e54ba6d951b0683e59cf1a68229f