teach RpcServer to do takeover
Summary:
We are working towards supporting graceful restart on NFS so that NFS users will
have this available when we start the migration.

This commit is teaching the Server at the core of the nfsd and mountd to
support takeoverShutdown and takeoverStartup. This means stopping processing
requests when we call takover shutdown and passing the sockets (both server
socket and connected sockets) up the stack.

For now we are only supporting takeover once everything has been initialized
and not during mount or unmount.

This means for the nfsd we will only care about the one connected socket which
is being used to communicate with the kernel. we will pass through the
shutdown promise for the channel. For the mountd we only care about the server
socket which we will return from the takeoverShutdown method itself.

Then we also add being able to initialize itself from a server socket or
a connected socket.

Reviewed By: xavierd

Differential Revision: D31802790

fbshipit-source-id: 06ef4da48cd864bfa20ea0fb98485792255d98cd