takeover stop for NFS mounts
Summary:
For restart for NFS we now have the pieces for the nfs channel to
perform takeover and for the takeover code to transfer the data.

This commit plumbs everything through so that we are able to stop and nfs
channel on takeover. this means extracting the connected socket and plubming it
to the takeover server.

Reviewed By: xavierd

Differential Revision: D31357367

fbshipit-source-id: 91973f7c149ae9209e85aa40111f41e7753db63f