enable takeover tests
Summary:
Now that graceful restart is supported, we can enable the integration tests!

These are all passing. The IO test is a little flakey, but the fuse version is
as well. So I haven't invested a lot of time in it.

One test is only compatible with version 1 and 2 of the takeover protocol
which is not compatible with takeover for nfs mounts. So that test will remain
disabled.

Reviewed By: xavierd

Differential Revision: D31738202

fbshipit-source-id: 0160310355138db59a8e671d3888b948e896eefa