fschannel: do not report the true used space
Summary:
We've seen several reports of users getting confused by the output of `df` and
thinking that EdenFS was using 100+ GBs of disk space. In reality, EdenFS is
simply reporting the stats from the underlying filesystem, and thus isn't a
true representation of the disk space usage.

To avoid this confusion, let's simply pretend that EdenFS doesn't use any disk
space. This isn't exactly true as the overlay and some internal datastructures
are using space already, but `eden du` is the right way to see these.

Reviewed By: fanzeyi

Differential Revision: D32536715

fbshipit-source-id: ec9b27084340d2e69cc5a14d6aa71def5fb4465d