edenapi: test land_stack method
Summary:
Add an integration test for the new EdenApi `land_stack` method.

We set up a commit graph on the server, perform a rebase and update the bookmark from the client using land_stack, and then pull down the changes and verify the client can observe the new commit graph.

Reviewed By: quark-zju

Differential Revision: D32362703

fbshipit-source-id: 94816f73d534a50521c966902e4704feb5abfc15