edenapi: implement ToWire & ToApi for HashMap
Summary:
Add `ToWire` implementations for `HashMap`, so that it can be used in edenapi types. Because JSON doesn't support non-string keys, I currently implement `WireMap` as `Vec<(K, V)>` rather than implementing `ToApi` for `HashMap` and serializating it directly.

This will be used in `LandStackResponse` to get determinstic ordering for the rebase updates in the `debugapi` pretty printing.

Reviewed By: quark-zju

Differential Revision: D32507125

fbshipit-source-id: 2f6ab203b6e1fd64228d76bbe96db1645cb50ab2