mononoke: derive hg changesets in parallel
Summary:
Unodes/fsnodes/skeleton manifests support new parallel derivation, let's add
support for hg changesets. The main difference is that commits with copy info
can't be derived with new parallel derivation. Rename source is part of the
rename destination filenode, and it's awkward to add this to new parallel
derivation. So i decided to derive commits with renames the old way just as we
do for merges.

There are a few things that we could optimize - for example, changed file
generation can be done in parallel. We can do it in later diffs if necessary.

Also renamed "file_changes" field to "stack_items" to make the name clearer

Reviewed By: farnz

Differential Revision: D31799384

fbshipit-source-id: 3775012e836d1a91b77cb0bae95abb66ce5ad215