mononoke: use CrossRepoSync constructor in create_commit_syncers
Summary:
Previously we were bypassing CrossRepoSync constructor, however it does a
necessary validation - it checks if these two repos can be used together at
all.

Let's fix it in this diff. Note that to do that I also de-asyncified
get_common_config() function since it didn't really have to be async, and in
order to use it in create_commit_syncers we need to deasyncify it since
create_commit_syncers is not async.

Reviewed By: mitrandir77

Differential Revision: D32495138

fbshipit-source-id: e5e458b7763c8f63ac7fd8b088bd6d1d1f2e490b