mononoke: use new parallel derivation in tests
Summary:
I'd like to enable new parallel derivation soonish, so it would be good to
first start using it in our integration tests.

And while running these tests I've found bugs in fsnodes and skeleton manifests! so this diff also fixes them. The problem was that we had two different hashmaps to track which commits are derived, and it was easy to confuse which hashmap should be used for fetch_unknown_parents function. Now there's only one hashmap, so it's much harder to get confused

Reviewed By: farnz

Differential Revision: D32665902

fbshipit-source-id: f528f64654ded230e12292c110f06cf9b02e9bef