hgmain: disable chg if stdin is not a tty but stdout is
Summary:
If stdout is a tty, then interactive is enabled, pager can start with stdio.
However, if the stdin is not a tty, then the user loses key bindings control to
the pager.

I also spent some time trying to get the `/dev/tty` passed from client to the
forked worker. The fd can be passed, but the worker cannot set that fd to
`/dev/tty`. POSIX does not provide a way. Linux has `ioctl(fd, TIOCSCTTY)`,
but when the terminal is owned by a different session, it requires
`CAP_SYS_ADMIN` to "steal" the ctty.

Maybe streampager can provide a `Pager::from_tty_fd` API. But for now let's
just disable chg in this case.

See also the next diff.

Reviewed By: DurhamG

Differential Revision: D32598519

fbshipit-source-id: cb7cf78aaf32f12fabca7b15eceabdea721e23dc